Added location reverse lookup using nominatim openstreetmap api (#21)
* Updates for mobile display

* Update for lint test

* resized filters for mobile display

* Changed map display defaults, Refactor file paths and import statements

* ignore for now to pass deploy checks

* Added location reverse lookup using nominatim openstreetmap api

* Changed map display defaults, Refactor file paths and import statements
