+++ src/styles/app.css
@@ -99,12 +99,25 @@
 
 .info-window-title {
   margin-top: 10px;
-  margin-bottom: 6px;
+  margin-bottom: 10px;
   text-align: center;
   text-overflow: "ellipsis";
   white-space: nowrap;
   overflow: hidden;
   width: auto;
+}
+.info-window-location {
+  margin-top: 2px;
+  margin-bottom: 6px;
+  text-align: start;
+  text-overflow: "ellipsis";
+  white-space: nowrap;
+  overflow: hidden;
+  width: auto;
+  margin-bottom: 10px;
+  /* font-weight: bold; */
+  /* line-height: 16px;
+  height: 33px; */
 }
 
 .info-window-image {
@@ -115,7 +128,7 @@
 
 .info-window-desc {
   width: stretch;
-  height: 87px;
+  height: 89px;
   margin: 0;
   line-height: 18px;
   overflow: scroll;
@@ -134,6 +147,7 @@
   text-overflow: ellipsis;
   vertical-align: auto;
   padding-top: 3px;
+  align-tracks: center;
 }
 
 .info-window-time-select {
